# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.PsiObservations:
    properties:
      content:
        type: string
      create_by:
        description: CreateBy almacena el nombre o identificador textual del creador.
        type: string
      create_by_id:
        description: |-
          CreateById es el UUID del usuario/administrador que creó el registro.
          Es un puntero para permitir valores nulos si la creación es automática por el sistema.
        type: string
      created_at:
        description: |-
          CreatedAt registra la fecha y hora exacta en que se creó el registro.
          GORM gestiona este campo automáticamente durante la inserción.
        type: string
      id:
        type: string
      psi_user_id:
        type: string
      update_by:
        description: UpdateBy almacena el nombre o identificador textual de la última
          persona en modificarlo.
        type: string
      update_by_id:
        description: UpdateById es el UUID del usuario/administrador que realizó la
          última actualización.
        type: string
      updated_at:
        description: |-
          UpdatedAt registra la última fecha y hora en que se modificó el registro.
          GORM actualiza este valor automáticamente en cada operación de guardado.
        type: string
    type: object
...
  request_structs.CreateObservacionesRequest:
    properties:
      content:
        description: |-
          Content es el texto libre de la observación.
          Es texto plano; la capa de servicio lo sanitiza con bluemonday antes de persistir.
        example: Nota interna de seguimiento.
        type: string
    required:
    - content
    type: object
...
  request_structs.UpdateObservacionesRequest:
    properties:
      content:
        description: Content es el nuevo texto de la observación (texto plano; se
          sanitiza en la API).
        example: Nota interna actualizada.
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend para la gestión del Colegio de Psicólogos de Carabobo.
  title: ColPsiCarabobo API
  version: "1.0"
paths:
...
  /admin/psi/{id}/observaciones:
    get:
      description: Retorna TODAS las observaciones internas de un psicólogo. Acceso
        exclusivo del personal administrativo autorizado; el psicólogo nunca ve sus
        propias observaciones.
      parameters:
      - description: UUID del Psicólogo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.PsiObservations'
            type: array
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permisos insuficientes
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Psicólogo no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar observaciones internas (Admin)
      tags:
      - Administración - Psicólogos
    post:
      consumes:
      - application/json
      description: Registra una nueva observación interna sobre un psicólogo. El contenido
        es texto plano y se sanitiza en la API.
      parameters:
      - description: UUID del Psicólogo
        in: path
        name: id
        required: true
        type: string
      - description: Contenido de la observación
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_structs.CreateObservacionesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Observación creada'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: ID inválido o contenido vacío'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Permisos insuficientes'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Psicólogo no encontrado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Crear observación interna (Admin)
      tags:
      - Administración - Psicólogos
  /admin/psi/{id}/observaciones/{entryId}:
    patch:
      consumes:
      - application/json
      description: Edita el contenido de una observación interna existente. El contenido
        es texto plano y se sanitiza en la API.
      parameters:
      - description: UUID del Psicólogo
        in: path
        name: id
        required: true
        type: string
      - description: UUID de la observación
        in: path
        name: entryId
        required: true
        type: string
      - description: Nuevo contenido de la observación
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_structs.UpdateObservacionesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Observación actualizada'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: ID inválido o contenido vacío'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Permisos insuficientes'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Observación no encontrada'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Editar observación interna (Admin)
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"